feat: move autoscaling helpers to `@crawlee/core`
Moves autoscaling-related single-consumer symbols (CPU/memory/cgroup
detection, ps-tree) from the public @crawlee/utils package into
@crawlee/core, since @crawlee/core is their only consumer. Removes the
now-unused CpuSample, MemoryInfo, getCurrentCpuTicksV2, getMemoryInfo,
isDocker, isContainerized, isLambda, and getCgroupsVersion exports from
@crawlee/utils.

Related to #3079.
