+++ docs/public-api/crawlee-utils.api.md
@@ -43,14 +43,6 @@ export const CLOUDFLARE_RETRY_CSS_SELECTORS: string[];
 export { Constructor }
 
 // @public
-export interface CpuSample {
-    // (undocumented)
-    containerUsage: number;
-    // (undocumented)
-    systemUsage: number;
-}
-
-// @public
 export function createRequestDebugInfo(request: Request_2, response?: IncomingMessage | Partial<BrowserResponseLike>, additionalFields?: Dictionary): Dictionary;
 
 export { Dictionary }
@@ -121,21 +113,6 @@ const FACEBOOK_REGEX: RegExp;
 // @public
 const FACEBOOK_REGEX_GLOBAL: RegExp;
 
-// @public
-export function getCgroupsVersion(forceReset?: boolean): Promise<"V1" | "V2" | null>;
-
-// @internal
-export function getCurrentCpuTicksV2(options?: {
-    containerized?: boolean;
-    logger?: CrawleeLogger;
-}): Promise<number>;
-
-// @internal
-export function getMemoryInfo(options?: {
-    containerized?: boolean;
-    logger?: CrawleeLogger;
-}): Promise<MemoryInfo>;
-
 // @public (undocumented)
 export function getObjectType(value: unknown): string;
 
@@ -157,17 +134,8 @@ export function isAsyncIterable<T>(value: unknown): value is AsyncIterable<T>;
 // @public
 export function isBuffer(value: unknown): value is Buffer | ArrayBuffer | ArrayBufferView;
 
-// @public
-export function isContainerized(): Promise<boolean>;
-
-// @public
-export function isDocker(forceReset?: boolean): Promise<boolean>;
-
 // @internal
 export function isIterable<T>(value: unknown): value is Iterable<T>;
-
-// @public (undocumented)
-export function isLambda(): boolean;
 
 // @public
 export function isStream(value: unknown): value is NodeJS.ReadableStream | ReadableStream;
@@ -180,15 +148,6 @@ const LINKEDIN_REGEX: RegExp;
 
 // @public
 const LINKEDIN_REGEX_GLOBAL: RegExp;
-
-// @public
-export interface MemoryInfo {
-    childProcessesBytes: number;
-    freeBytes: number;
-    mainProcessBytes: number;
-    totalBytes: number;
-    usedBytes: number;
-}
 
 // @public
 export function mergeAsyncIterables<T>(...iterables: AsyncIterable<T>[]): AsyncIterable<T>;
