+++ packages/core/src/system-info/ps-tree.ts
