+++ test/core/autoscaling/fixtures/child.js
