+++ test/core/autoscaling/fixtures/parent.js
@@ -1,5 +1,5 @@
 import { exec } from 'node:child_process';
 
 for (let count = 1; count < 10; count++) {
-    exec('node ./test/utils/fixtures/child.js');
+    exec('node ./test/core/autoscaling/fixtures/child.js');
 }
