+++ test/core/autoscaling/psTree.test.ts
@@ -1,7 +1,7 @@
 import { exec } from 'node:child_process';
 import path from 'node:path';
 
-import { psTree } from '../../packages/utils/src/internals/system-info/ps-tree.js';
+import { psTree } from '../../../packages/core/src/system-info/ps-tree.js';
 
 const scripts = {
     parent: path.join(import.meta.dirname, 'fixtures', 'parent.js'),
